+++ src/doomstat.h
@@ -238,7 +238,7 @@ extern boolean coop_spawns;
 extern int snd_SfxVolume;      // maximum volume for sound
 extern int snd_MusicVolume;    // maximum volume for music
 
-// [Cherry] Mute inactive window
+// [Cherry] Mute Inactive Window feature from International Doom
 
 extern int sfx_volume, music_volume;
 extern boolean mute_inactive;